业务提交

+++ src/main/java/com/mybatis/spring/common/IErrorCode.java
@@ -5,7 +5,7 @@ package com.mybatis.spring.common;
  * Created by macro on 2019/4/19.
  */
 public interface IErrorCode {
-    long getCode();
+    Integer getCode();
 
     String getMessage();
 }
